+++ R/edit-data-utils.R
@@ -13,6 +13,8 @@
 #' @param colnames `data.frame` column names
 #' @param var_edit vector of `character` which allows to choose the editable columns
 #' @param var_mandatory vector of `character` which allows to choose obligatory fields to fill
+#' @param modal_size `character` which allows to choose the size of the modalDialog. One of "s" for small, "m" (the default) for medium, "l" for large, or "xl" for extra large.
+#' @param modal_easy_close `boolean` If TRUE, modalDialog can be dismissed by clicking outside the dialog box, or be pressing the Escape key. If FALSE (the default), modalDialog can't be dismissed in those ways; instead it must be dismissed by clicking on a modalButton(), or from a call to removeModal() on the server.
 #' @param session The `session` object passed to function given to shinyServer
 #'
 #' @importFrom shiny showModal modalDialog actionButton
@@ -25,18 +27,17 @@ edit_modal <- function(default = list(),
                        id_validate = "add_row",
                        title = i18n("Add a row"),
                        data,
-                       colnames = names(data),
-                       var_edit,
-                       var_mandatory,
+                       var_edit = NULL,
+                       var_mandatory = NULL,
+                       var_labels = colnames(data),
+                       modal_size = "m",
+                       modal_easy_close = FALSE,
+                       n_column = 1,
                        session = getDefaultReactiveDomain()) {
   ns <- session$ns
 
-  if (identical(x = var_edit, y = character(0)) | identical(x = var_edit, y = NULL)) {
-    data <- data
-    position_var_edit <- seq_len(ncol(data))
-  } else {
+  if (length(var_edit) > 0) {
     data <- data[, ..var_edit]
-    position_var_edit <- as.numeric(gsub("col_", "", var_edit))
   }
 
   showModal(modalDialog(
@@ -52,19 +53,21 @@ edit_modal <- function(default = list(),
       )
     ),
     footer = NULL,
-    size = "m",
-    easyClose = TRUE,
+    size = modal_size,
+    easyClose = modal_easy_close,
     edit_input_form(
       default = default,
       data = data,
-      colnames = colnames,
       var_mandatory = var_mandatory,
-      position_var_edit =  position_var_edit,
+      var_labels = var_labels,
+      n_column = n_column,
       session = session
     ),
     actionButton(
       inputId = ns(id_validate),
-      label = i18n("Validate the entry"),
+      label = tagList(
+        ph("floppy-disk"), i18n("Save")
+      ),
       class = "btn-outline-primary float-end"
     )
   ))
@@ -85,55 +88,89 @@ edit_modal <- function(default = list(),
 #' @importFrom shiny numericInput textInput
 #' @importFrom shinyWidgets virtualSelectInput prettyCheckbox airDatepickerInput
 #' @importFrom htmltools tagList tags
+#' @importFrom rlang is_vector is_named
 #'
 #' @return different shiny widgets with edited columns according to their respective class
 #' @noRd
 #'
-edit_input_form <- function(default = list(), data, colnames, var_mandatory, position_var_edit, session = getDefaultReactiveDomain()) {
+edit_input_form <- function(default = list(),
+                            data,
+                            var_mandatory = NULL,
+                            var_labels = colnames(data),
+                            n_column = 1,
+                            session = getDefaultReactiveDomain()) {
 
   ns <- session$ns
 
-  tagList(
-    lapply(
-      X = seq_len(ncol(data)),
-      FUN = function(i) {
-        variable_id <- colnames(data)[i]
-        variable_name <- colnames[position_var_edit[i]]
-        variable <- data[[i]]
+  if (is_vector(var_labels) & !is_named(var_labels)) {
+    var_labels <- setNames(var_labels, unlist(var_labels))
+  }
 
-        if (variable_name %in% var_mandatory) {
-          label <- tagList(variable_name, tags$span(HTML("&#42;"), class = "asterisk", style = "color: red;"), " : ")
-        } else {
-          label <- paste0(variable_name, " : ")
-        }
+  widgets <- lapply(
+    X = seq_len(ncol(data)),
+    FUN = function(i) {
+      variable_id <- colnames(data)[i]
+      variable_label <- var_labels[which(names(var_labels) == variable_id)]
+      if (length(variable_label) < 1)
+        variable_label <- variable_id
+      variable <- data[[i]]
 
-        if (isTRUE((inherits(x = variable, what = "numeric")))) {
-          numericInput(
+      suffix <- if (isTRUE((inherits(variable, "logical")))) "" else " : "
+      if (variable_id %in% var_mandatory) {
+        label <- tagList(
+          variable_label,
+          tags$span(HTML("&#42;"), class = "asterisk", style = "color: red;"), suffix
+        )
+      } else {
+        label <- paste0(variable_label, suffix)
+      }
+
+      if (isTRUE(inherits(variable, c("numeric", "integer")))) {
+        opts <- getOption("datamods.edit.input.numeric", list())
+        opts <- modifyList(
+          x = opts,
+          val = list(
             inputId = ns(variable_id),
             label = label,
             value = default[[variable_id]] %||% NA_real_,
             width = "100%"
           )
-        } else if (isTRUE((inherits(x = variable, what = "factor")))) {
-          virtualSelectInput(
+        )
+        do.call(numericInput, opts)
+      } else if (isTRUE((inherits(variable, "factor")))) {
+        opts <- getOption("datamods.edit.input.factor", list())
+        opts <- modifyList(
+          x = opts,
+          val = list(
             inputId = ns(variable_id),
             label = label,
-            choices = unique(variable),
+            choices = sort(unique(c(as.character(variable), levels(variable)))),
             selected = default[[variable_id]] %||% "",
             width = "100%",
             allowNewOption = TRUE,
             autoSelectFirstOption = FALSE,
-            placeholder = i18n("Select")
+            placeholder = i18n("Select"),
+            zIndex = 999
           )
-        } else if (isTRUE((inherits(x = variable, what = "character")))) {
-          textInput(
+        )
+        do.call(virtualSelectInput, opts)
+      } else if (isTRUE((inherits(variable, "character")))) {
+        opts <- getOption("datamods.edit.input.character", list())
+        opts <- modifyList(
+          x = opts,
+          val = list(
             inputId = ns(variable_id),
             label = label,
             value = default[[variable_id]] %||% "",
             width = "100%"
           )
-        } else if (isTRUE((inherits(x = variable, what = "logical")))) {
-          prettyCheckbox(
+        )
+        do.call(textInput, opts)
+      } else if (isTRUE((inherits(variable, "logical")))) {
+        opts <- getOption("datamods.edit.input.logical", list())
+        opts <- modifyList(
+          x = opts,
+          val = list(
             inputId = ns(variable_id),
             label = label,
             value = default[[variable_id]] %||% FALSE,
@@ -141,17 +178,49 @@ edit_input_form <- function(default = list(), data, colnames, var_mandatory, pos
             status = "primary",
             width = "100%"
           )
-        } else if (isTRUE((inherits(x = variable, what = "Date")))) {
-          airDatepickerInput(
+        )
+        do.call(prettyCheckbox, opts)
+      } else if (isTRUE((inherits(variable, "Date")))) {
+        opts <- getOption("datamods.edit.input.Date", list())
+        opts <- modifyList(
+          x = opts,
+          val = list(
             inputId = ns(variable_id),
             label = label,
             value = default[[variable_id]] %||% Sys.Date(),
-            inline = TRUE,
             width = "100%"
           )
-        } else {
-          return(NULL)
-        }
+        )
+        do.call(airDatepickerInput, opts)
+      } else if (isTRUE((inherits(variable, c("POSIXct", "POSIXt"))))) {
+        opts <- getOption("datamods.edit.input.POSIXt", list())
+        opts <- modifyList(
+          x = opts,
+          val = list(
+            inputId = ns(variable_id),
+            label = label,
+            value = default[[variable_id]] %||% Sys.time(),
+            timepicker = TRUE,
+            width = "100%"
+          )
+        )
+        do.call(airDatepickerInput, opts)
+      } else {
+        return(NULL)
+      }
+    }
+  )
+  fluidRow(
+    lapply(
+      X = split(
+        x = seq_along(widgets),
+        f = rep(seq_len(n_column), each = ceiling(length(widgets)/n_column))[seq_along(widgets)]
+      ),
+      FUN = function(i) {
+        column(
+          width = 12 / n_column,
+          widgets[i]
+        )
       }
     )
   )
@@ -164,34 +233,93 @@ edit_input_form <- function(default = list(), data, colnames, var_mandatory, pos
 #'
 #' @param data `data.frame` to use
 #' @param colnames `data.frame` column names
+#' @param reactable_options `list` allowing you to add reactable options
 #'
 #' @return the `data.frame` in reactable format
 #' @noRd
 #'
 #' @importFrom reactable reactable colDef
-#'
-table_display <- function(data, colnames = NULL) {
-  cols <- list()
-  for (i in seq_along(data)) {
-    cols[[names(data)[i]]] <- colDef(name = colnames[i])
+#' @importFrom data.table copy setnames
+table_display <- function(data, colnames = NULL, reactable_options = NULL) {
+
+  data <- copy(data)
+  if (!is.null(colnames)) {
+    setnames(data, old = seq_along(colnames), new = colnames)
   }
+
+  cols <- reactable_options$columns %||% list()
   if (all(is.na(data$.datamods_edit_update))) {
-    cols$.datamods_edit_update = colDef(show = FALSE)
+    cols$.datamods_edit_update <- colDef(show = FALSE)
   } else {
-    cols$.datamods_edit_update = col_def_update()
+    cols$.datamods_edit_update <- col_def_update()
   }
 
   if (all(is.na(data$.datamods_edit_delete))) {
-    cols$.datamods_edit_delete = colDef(show = FALSE)
+    cols$.datamods_edit_delete <- colDef(show = FALSE)
   } else {
-    cols$.datamods_edit_delete = col_def_delete()
+    cols$.datamods_edit_delete <- col_def_delete()
   }
 
   cols$.datamods_id <- colDef(show = FALSE)
-  reactable(
-    data = data,
-    columns = cols
-  )
+
+  if (is.null(reactable_options))
+    reactable_options <- list()
+  reactable_options <- reactable_options
+  reactable_options$data <- data
+  reactable_options$columns <- cols
+
+  rlang::exec(reactable::reactable, !!!reactable_options)
+}
+
+#' @importFrom reactable updateReactable getReactableState
+#' @importFrom data.table copy setnames
+update_table <- function(data, colnames) {
+  data <- copy(data)
+  setnames(data, old = seq_along(colnames), new = colnames)
+  page <- getReactableState(outputId = "table", name = "page")
+  updateReactable("table", data = data, page = page)
+  return(data)
+}
+
+format_edit_data <- function(data, colnames, internal_colnames = NULL) {
+  data <- as.data.table(data)
+  vars_datamods_edit <- intersect(c(".datamods_id", ".datamods_edit_update", ".datamods_edit_delete"), names(data))
+  data <- data[, -..vars_datamods_edit]
+  if (is.null(internal_colnames))
+    internal_colnames <- seq_along(colnames)
+  setnames(data, old = internal_colnames, new = colnames, skip_absent = TRUE)
+  data[]
+}
+
+rename_edit <- function(data, var_labels) {
+  for(i in seq_along(names(var_labels))) {
+    names(data)[names(data) == names(var_labels)[i]] <- var_labels[[i]]
+  }
+  data
+}
+
+
+#' @importFrom rlang set_names is_null is_list is_named
+get_variables_labels <- function(labels, column_names, internal_names) {
+  if (is_null(labels)) {
+    labels <- column_names
+  } else {
+    if (!is_list(labels)) {
+      stopifnot(
+        "If `var_labels` is an unnamed vector, it must have same length as `colnames(data)`" = length(labels) == length(column_names)
+      )
+      labels <- set_names(as.list(labels), column_names)
+    }
+    stopifnot(
+      "`var_labels` must be a named list" = is_named(labels)
+    )
+    names(labels) <- internal_names[match(names(labels), column_names)]
+    labels <- modifyList(
+      x = set_names(as.list(column_names), internal_names),
+      val = labels
+    )
+  }
+  return(labels)
 }
 
 
@@ -240,7 +368,7 @@ btn_update <- function(inputId) {
         title = i18n("Click to edit"),
         ph("pencil-simple-line", height = "1.2em")
       )
-   )
+    )
 
   }
 }
@@ -290,7 +418,7 @@ btn_delete <- function(inputId) {
         title = i18n("Click to delete"),
         ph("x", height = "1.2em")
       )
-   )
+    )
   }
 }
 
@@ -345,35 +473,44 @@ confirmation_window <- function(inputId, ..., title = NULL) {
 
 
 #' @importFrom shinybusy notify_failure notify_success notify_info notify_warning
-notification_failure <- function(title, text) {
-  shinybusy::notify_failure(
-    title = title,
-    text = text,
-    position = "center-top",
-    clickToClose = TRUE
-  )
+notification_failure <- function(title, text, use_notify = TRUE) {
+  if (isTRUE(use_notify)) {
+    shinybusy::notify_failure(
+      title = title,
+      text = text,
+      position = "center-top",
+      clickToClose = TRUE
+    )
+  }
 }
-notification_warning <- function(title, text) {
-  shinybusy::notify_warning(
-    title = title,
-    text = text,
-    position = "center-top",
-    clickToClose = TRUE
-  )
+notification_warning <- function(title, text, use_notify = TRUE) {
+  if (isTRUE(use_notify)) {
+    shinybusy::notify_warning(
+      title = title,
+      text = text,
+      position = "center-top",
+      clickToClose = TRUE
+    )
+  }
 }
-notification_success <- function(title, text) {
-  shinybusy::notify_success(
-    title = title,
-    text = text,
-    position = "center-top",
-    clickToClose = TRUE
-  )
+notification_success <- function(title, text, use_notify = TRUE) {
+  if (isTRUE(use_notify)) {
+    shinybusy::notify_success(
+      title = title,
+      text = text,
+      position = "center-top",
+      clickToClose = TRUE
+    )
+  }
 }
-notification_info <- function(title, text) {
-  shinybusy::notify_info(
-    title = title,
-    text = text,
-    position = "center-top",
-    clickToClose = TRUE
-  )
+notification_info <- function(title, text, use_notify = TRUE) {
+  if (isTRUE(use_notify)) {
+    shinybusy::notify_info(
+      title = title,
+      text = text,
+      position = "center-top",
+      clickToClose = TRUE
+    )
+  }
 }
+