+++ NEWS.md
@@ -1,3 +1,12 @@
+# datamods 1.4.3
+
+* `edit_data_server()` : added the ability to specify callbacks functions to be executed before performing an action on the table (add, update or delete).
+* `edit_data_server()` : pass reactable option + selection to the table [#82](https://github.com/dreamRs/datamods/pull/82)
+* `edit-data` module : use factor levels and sort theme in edit input forma for factors (sorting also applies for characters), thanks to [@Felixmil](https://github.com/Felixmil).
+* `import-file` module : allow to specify string used to identify `NA`, thanks to [@DrFabach](https://github.com/DrFabach).
+* `filter_data_server()` : argument `drop_ids` can now be set via option `datamods.filter.drop_ids` and can be a list like `list(p = 0.9, n = 50)` to specify threshold values to remove IDs columns.
+
+
 # datamods 1.4.2
 
 * i18n: japanese translations added, thanks to [@nissinbo](https://github.com/nissinbo).
@@ -12,7 +21,7 @@
 * Fixed displaying variable class in View tab (fix [#64](https://github.com/dreamRs/datamods/issues/64)).
 * `select_group_server()` : fix update inputs when `multiple = FALSE`.
 * `filter_data_server()` : sorting choices in select menus (select, picker and virtual) (fix [#66](https://github.com/dreamRs/datamods/issues/64))).
-* `filter_data_server()` : don't use `<`/`>` for empty field to not confuse to an HTMl tag (fix [#65](https://github.com/dreamRs/datamods/issues/65))).
+* `filter_data_server()` : don't use `<`/`>` for empty field to not confuse to an HTML tag (fix [#65](https://github.com/dreamRs/datamods/issues/65))).
 
 
 # datamods 1.4.0
@@ -57,7 +66,7 @@
 * `import_copypaste_server()`: new argument `fread_args` to pass arguments to `data.table::fread`.
 * i18n: chinese translations added, thanks to [@xmusphlkg](https://github.com/xmusphlkg).
 * i18n: spanish translations added, thanks to [@dnldelarosa](https://github.com/dnldelarosa).
-* i18n: german translations added, thanks to [@SteEcker](https://github.com/SteEcker) and [@joerghenkebuero](https://github.com/joerghenkebuero).
+* i18n: german translations added, thanks to [@SteEcker](https://github.com/SteEcker) and joerghenkebuero.